+++ lib/connectanum.dart
@@ -28,4 +28,4 @@ export 'src/transport/abstract_transport.dart';
 export 'src/transport/websocket/websocket_transport_serialization.dart';
 export 'src/transport/websocket/websocket_transport_none.dart'
     if (dart.library.io) 'src/transport/websocket/websocket_transport_io.dart' // dart:io implementation
-    if (dart.library.html) 'src/transport/websocket/websocket_transport_web.dart'; // package:web/web.dart implementation
+    if (dart.library.js_interop) 'src/transport/websocket/websocket_transport_web.dart'; // package:web/web.dart implementation